requirements.txt
numpy
pandas
matplotlib
seaborn
warnings
sklearn.model_selection
sklearn.preprocessing MinMaxScaler
sklearn.linear_model LogisticRegression
sklearn.metrics  *

